patch: async email sending via graphql
