+++ api/graphql/mappers/email.go
@@ -4,6 +4,7 @@ import (
 	"github.com/customeros/mailstack/api/graphql/graphql_model"
 	"github.com/customeros/mailstack/internal/enum"
 	"github.com/customeros/mailstack/internal/models"
+	"github.com/customeros/mailstack/internal/utils"
 )
 
 func MapGormEmailToGraph(email *models.Email) *graphql_model.EmailMessage {
@@ -24,18 +25,18 @@ func MapGormEmailToGraph(email *models.Email) *graphql_model.EmailMessage {
 
 func MapGraphEmailInputToGorm(email *graphql_model.EmailInput) *models.Email {
 	return &models.Email{
-		MailboxID:    *email.MailboxID,
+		MailboxID:    utils.GetOrDefault(email.MailboxID, ""),
 		Direction:    enum.EmailDirectionOutbound,
-		FromAddress:  email.FromAddresss,
-		FromName:     *email.FromName,
+		FromAddress:  email.FromAddress,
+		FromName:     utils.GetOrDefault(email.FromName, ""),
 		ToAddresses:  email.ToAddresses,
 		CcAddresses:  email.CcAddresses,
 		BccAddresses: email.BccAddresses,
-		ReplyTo:      *email.ReplyTo,
+		ReplyTo:      utils.GetOrDefault(email.ReplyTo, ""),
 		Subject:      email.Subject,
-		BodyText:     *email.Body.Text,
-		BodyHTML:     *email.Body.HTML,
+		BodyText:     utils.GetOrDefault(email.Body.Text, ""),
+		BodyHTML:     utils.GetOrDefault(email.Body.HTML, ""),
 		ScheduledFor: email.ScheduleFor,
-		TrackClicks:  *email.TrackClicks,
+		TrackClicks:  utils.GetOrDefault(email.TrackClicks, false),
 	}
 }
