+++ deployments/docker-compose.yaml
@@ -28,6 +28,7 @@ services:
     ports:
       - "16686:16686"
       - "6831:6831/udp"
+      - "6831:6831/tcp"
       - "4317:4317"
       - "4318:4318"
     environment:
