+++ dto/send_email.go
@@ -1,0 +1,7 @@
+package dto
+
+import "github.com/customeros/mailstack/internal/models"
+
+type SendEmail struct {
+	Email *models.Email
+}
