--- errors/errors.go
@@ -1,8 +1,0 @@
-package mailstack_errors
-
-import "errors"
-
-var (
-	ErrTenantNotSet = errors.New("tenant not set on context")
-	ErrUserIDNotSet = errors.New("userId not set on context")
-)
